+++ src/ContextInterface2.php
@@ -29,4 +29,18 @@ interface ContextInterface2 extends ContextInterface
      * @return SQLSyntaxHelpersInterface the SQL syntax helpers corresponding to the SQL type returned by getSqlType()
      */
     public function getSqlSyntaxHelpers() : SQLSyntaxHelpersInterface;
+
+    /**
+     * Convert the given path, representing a PHP class implementing a factory,
+     * to the corresponding CustomClassFileInterface object.
+     *
+     * The path can be a system file path, or an URI, or any other structured
+     * name representing the class file. The path type depends on the framework
+     * or the environment where JelixDao is used.
+     *
+     * @param string $path
+     *
+     * @return CustomClassFileInterface
+     */
+    public function resolveCustomFactoryClassPath($path);
 }